refactor: dto를 model로 옮겨서 model이 view에 의존하지 않게 수정
