+++ src/main/java/blackjack/controller/BlackjackGame.java
@@ -2,16 +2,16 @@ package blackjack.controller;
 
 import blackjack.model.betting.Betting;
 import blackjack.model.betting.BettingMoney;
+import blackjack.model.betting.PlayerBettingProfitOutcome;
 import blackjack.model.cardgenerator.CardGenerator;
 import blackjack.model.cardgenerator.RandomCardGenerator;
 import blackjack.model.dealer.Dealer;
+import blackjack.model.dealer.DealerFinalCardsOutcome;
 import blackjack.model.player.Player;
+import blackjack.model.player.PlayerFinalCardsOutcome;
 import blackjack.model.player.Players;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
-import blackjack.view.dto.DealerFinalCardsOutcome;
-import blackjack.view.dto.PlayerBettingProfitOutcome;
-import blackjack.view.dto.PlayerFinalCardsOutcome;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
