+++ src/main/java/blackjack/model/betting/PlayerBettingProfitOutcome.java
@@ -1,4 +1,4 @@
-package blackjack.view.dto;
+package blackjack.model.betting;
 
 public record PlayerBettingProfitOutcome(String name, int profit) {
 }
