+++ src/main/java/blackjack/model/dealer/Dealer.java
@@ -1,9 +1,9 @@
 package blackjack.model.dealer;
 
+import blackjack.model.betting.PlayerBettingProfitOutcome;
 import blackjack.model.card.Card;
 import blackjack.model.card.Cards;
 import blackjack.model.cardgenerator.CardGenerator;
-import blackjack.view.dto.PlayerBettingProfitOutcome;
 import java.util.List;
 
 public class Dealer {
