+++ src/main/java/blackjack/model/player/PlayerFinalCardsOutcome.java
@@ -1,7 +1,6 @@
-package blackjack.view.dto;
+package blackjack.model.player;
 
 import blackjack.model.card.Card;
-import blackjack.model.player.Player;
 import java.util.List;
 
 public record PlayerFinalCardsOutcome(String name, List<Card> cards, int totalScore) {
