+++ src/test/java/blackjack/model/dealer/DealerTest.java
@@ -2,11 +2,11 @@ package blackjack.model.dealer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.betting.PlayerBettingProfitOutcome;
 import blackjack.model.card.Card;
 import blackjack.model.card.Denomination;
 import blackjack.model.card.Suit;
 import blackjack.model.cardgenerator.SequentialCardGenerator;
-import blackjack.view.dto.PlayerBettingProfitOutcome;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
